+++ src/Dap.Archive/WebSocket/TextRecorder.fs
@@ -4,6 +4,7 @@ module Dap.Archive.WebSocket.TextRecorder
 open FSharp.Control.Tasks.V2
 
 open Dap.Prelude
+open Dap.Context
 open Dap.Platform
 open Dap.Remote
 open Dap.Archive